requirements.txt
requests
python-telegram-bot
cloudscraper